Fixed issue when packaging
The default project package was a uberjar, which was including 3rd party libraries code. That's not correct since can have impact in licenses. So the uberjar is provided only as an option '-Puberjar' and default packaging and deploying is an standalone module with dependencies.

Also updated the maven plugins versions