Explicitly flush per table mutations in the HBase applier, this will prevent background buffering of the mutations. This patch is intended to reduce the potential for data loss in the event that transactions are marked as flushed when in reality they are actually buffered waiting for timeout/size limit to be hit.
Note that with this modification it becomes much more important to ensure a proper value is set for any of the configuration
options related to buffer timeout (ie:  'applier.hbase.buffer.time.limit', 'applier.hbase.buffer.time.maximum', 'applier.hbase.buffer.time.minimum' )
as these will now have a direct and immediate impact on the performance of the application. The most important thing to note here is that the timeout
is measured in milliseconds.