First draft of implementing an option to allow the replicator to start up even if it has an invalid checkpoint.
Thoughts for next commit: should there be a separate configuration variable to enable/disable this behaviour
in the event that there is a checkpoint present, but it is invalid? Right now this only works without a checkpoint.

Also should this branch be used to add the functionality to bootstrap a "current" checkpoint in order to facilitate
starting a fresh replicator without replaying the stored binlogs?